move zn_textbox
